Rename chunks.ts to planLowResPrefetch.ts
